Realizado a refatoração

+++ src/br/com/samuel/calculadorajuroscompostos/repository/CalculadoraRepository.java
@@ -2,7 +2,7 @@ package br.com.samuel.calculadorajuroscompostos.repository;
 
 public class CalculadoraRepository {
 
-    public double calcularMontante(double capitalInicial, double taxa, int tempo) {
+    public double calcularMontante(double capitalInicial, double taxa, double tempo) {
         double taxaPorCento = taxa / 100.0;
         double somaTaxa = (1.0 + taxaPorCento);
         double somaTaxaElevado = Math.pow(somaTaxa, tempo);
